+++ crypto/src/test/java/org/cryptimeleon/incentive/crypto/IssueJoinCryptoTest.java
@@ -17,7 +17,6 @@ import java.util.logging.Logger;
 public class IssueJoinCryptoTest
 {
     Logger logger = Logger.getLogger(Logger.GLOBAL_LOGGER_NAME);
-
     /**
      * generates a join request, responds to it and handles the response (i.e. a full run of the Issue-Join protocol).
      * Test case simulating a correct use of the protocol.
@@ -29,7 +28,6 @@ public class IssueJoinCryptoTest
         logger.info("Starting unit test of the Issue-Join protocol.");
 
         logger.info("Setting up the incentive system and generating keys.");
-
         // create incentive system instance with fresh pp
         var incSys = new IncentiveSystem(IncentiveSystem.setup(128, Setup.BilinearGroupChoice.Debug), null);
 
@@ -43,9 +41,8 @@ public class IssueJoinCryptoTest
 
         // create join request
         var testRequest = incSys.generateJoinRequest(
-            incSys.getPp(),
-            pkp.getPk(),
-            ukp
+                pkp.getPk(),
+                ukp
         );
 
         logger.info("Testing join request serialization.");
@@ -58,7 +55,7 @@ public class IssueJoinCryptoTest
         logger.info("Generating a join response.");
 
         // pass join request to issue logic, generate join response
-        var testResponse = incSys.generateJoinRequestResponse(incSys.getPp(), pkp, ukp.getPk().getUpk(), deserializedRequest);
+        var testResponse = incSys.generateJoinRequestResponse(pkp, ukp.getPk().getUpk(), deserializedRequest);
 
         logger.info("Testing join response serialization.");
 
@@ -69,13 +66,13 @@ public class IssueJoinCryptoTest
         logger.info("Handling join request and updating token.");
 
         // pass join response to second part of join logic, generate join output
-        var testOutput = incSys.handleJoinRequestResponse(incSys.getPp(), pkp.getPk(), ukp, testRequest, deserializedResponse);
+        var testOutput = incSys.handleJoinRequestResponse(pkp.getPk(), ukp, testRequest, deserializedResponse);
 
         logger.info("Checking integrity of token.");
 
         // check output token for sanity (certficate valid, zero points)
         SPSEQSignatureScheme usedSpsEq = incSys.getPp().getSpsEq();
         Assertions.assertTrue(usedSpsEq.verify(pkp.getPk().getPkSpsEq(), testOutput.getSignature(), testOutput.getCommitment0(), testOutput.getCommitment1()));
-        Assertions.assertTrue(testOutput.getPoints().asInteger().compareTo(BigInteger.ZERO) == 0);
+        Assertions.assertEquals(testOutput.getPoints().asInteger().compareTo(BigInteger.ZERO), 0);
     }
 }