+++ services/basket/src/main/java/org/cryptimeleon/incentive/basket/exceptions/BasketServiceException.java
@@ -1,7 +1,0 @@
-package org.cryptimeleon.incentive.basket.exceptions;
-
-public class BasketServiceException extends Exception {
-    public BasketServiceException(String message) {
-        super(message);
-    }
-}