+++ services/basket/src/main/java/org/cryptimeleon/incentive/basket/exceptions/BasketUnauthorizedException.java
@@ -1,7 +1,0 @@
-package org.cryptimeleon.incentive.basket.exceptions;
-
-public class BasketUnauthorizedException extends BasketServiceException {
-    public BasketUnauthorizedException(String message) {
-        super(message);
-    }
-}