+++ services/basket/src/main/java/org/cryptimeleon/incentive/services/basket/BasketApplication.java
@@ -1,0 +1,11 @@
+package org.cryptimeleon.incentive.services.basket;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BasketApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(BasketApplication.class, args);
+    }
+}