+++ services/basket/src/main/java/org/cryptimeleon/incentive/services/basket/exceptions/BasketServiceException.java
@@ -1,0 +1,7 @@
+package org.cryptimeleon.incentive.services.basket.exceptions;
+
+public class BasketServiceException extends Exception {
+    public BasketServiceException(String message) {
+        super(message);
+    }
+}