+++ services/client/src/main/java/org/cryptimeleon/incentive/client/IssueClient.java
@@ -3,10 +3,13 @@ package org.cryptimeleon.incentive.client;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Mono;
 
-
+/**
+ * Client calls for info service.
+ * Can be used for testing and prototyping.
+ */
 public class IssueClient {
 
-    /*
+    /**
      * Webclient configured with the url of the issue service
      */
     private WebClient issueClient;
@@ -15,7 +18,7 @@ public class IssueClient {
         this.issueClient = WebClientHelper.buildWebClient(issueServiceUrl);
     }
 
-    /*
+    /**
      * Sends an request to the / endpoint which is configured to return the name of the service
      * This can be used to test whether a service is alive and reachable under some url
      */
@@ -26,4 +29,19 @@ public class IssueClient {
                 .bodyToMono(String.class);
     }
 
+    /**
+     * Creates a join request.
+     *
+     * @param serializedUserPublicKey the serialized public key of the user
+     * @param serializedJoinRequest   the serialized join request
+     * @return mono of the server's answer
+     */
+    public Mono<String> sendJoinRequest(String serializedJoinRequest, String serializedUserPublicKey) {
+        return issueClient.get()
+                .uri(uriBuilder -> uriBuilder.path("/issue").build())
+                .header("public-key", serializedUserPublicKey)
+                .header("join-request", serializedJoinRequest)
+                .retrieve()
+                .bodyToMono(String.class);
+    }
 }