+++ services/issue/src/main/java/org/cryptimeleon/incentive/services/issue/IssueController.java
@@ -3,40 +3,41 @@ package org.cryptimeleon.incentive.services.issue;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
-import org.cryptimeleon.incentive.services.issue.model.IssueResponse;
-import org.cryptimeleon.incentive.services.issue.model.JoinRequest;
+import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
+@AllArgsConstructor
 public class IssueController {
 
-    /*
+    private IssueService issueService;
+
+    /**
      * Endpoint for alive testing etc.
      */
     @GetMapping("/")
     public ResponseEntity<String> test() {
-        return new ResponseEntity<>("Issue Service", HttpStatus.OK);
+        return new ResponseEntity<>("Hello from Issue service!", HttpStatus.OK);
     }
 
     @GetMapping("/issue")
-    @ApiOperation(value = "Issuing protocol", notes = "Issue a new incentive token.", response = IssueResponse.class)
+    @ApiOperation(value = "Issuing protocol", notes = "Issue a new incentive token.", response = String.class)
     @ApiResponses({
-            @ApiResponse(code = 200, message = "Success", response = IssueResponse.class),
+            @ApiResponse(code = 200, message = "Success", response = String.class),
             @ApiResponse(code = 403, message = "Invalid Issuing Request", response = String.class)
     })
-    public ResponseEntity<IssueResponse> greeting(@Validated JoinRequest request) {
-        return new ResponseEntity<>(new IssueResponse(request.getId(), "Some serialized response"), HttpStatus.OK);
+    public ResponseEntity<String> performIssueJoin(
+            @RequestHeader(value = "join-request") String serializedJoinRequest,
+            @RequestHeader(value = "public-key") String serializedUserPublicKey
+    ) {
+        return new ResponseEntity<>(issueService.runIssueJoinProtocol(serializedJoinRequest, serializedUserPublicKey), HttpStatus.OK);
     }
 
     @ResponseStatus(HttpStatus.FORBIDDEN)
-    @ExceptionHandler(IncentiveException.class)
-    public String handleIncentiveException(IncentiveException ex) {
-        return "An incentive exception occurred!";
+    @ExceptionHandler(RuntimeException.class)
+    public String handleIncentiveException(RuntimeException ex) {
+        return "An exception occurred!";
     }
 }